+++ packages/terra-action-header/CHANGELOG.md
@@ -3,7 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.25.0 - (August 14, 2019)
+------------------
 ### Changed
+* Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts
 
 2.24.0 - (July 30, 2019)