+++ packages/terra-badge/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+3.20.0 - (August 14, 2019)
+------------------
 ### Changed
 * updated package.json test scripts