+++ packages/terra-breakpoints/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.15.0 - (August 14, 2019)
+------------------
 ### Changed
 * updated package.json test scripts