+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -3,8 +3,15 @@ Changelog
 
 Unreleased
 ----------
+
+3.22.0 - (August 14, 2019)
+------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts
+
+### Updated
+* updated jest snapshots
 
 3.21.0 - (July 30, 2019)
 ------------------
@@ -34,6 +41,7 @@ Unreleased
 3.16.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports
 * Updated Docs around the requirement of terra-base and aggregate-translations.