+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -2,7 +2,12 @@ ChangeLog
 =========
 
 Unreleased
------------------
+----------
+
+1.1.0 - (August 14, 2019)
+-------------------------
+### Changed
+* Minor dependency updates
 
 1.0.0 - (July 30, 2019)
 ------------------