+++ packages/terra-form-checkbox/CHANGELOG.md
@@ -3,11 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+
+3.22.0 - (August 14, 2019)
+------------------
 ### Changed
+* Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts
-
-
-### Changed
 * Update tests for dev-site v6
 
 3.21.0 - (July 30, 2019)
@@ -43,6 +44,7 @@ Unreleased
 3.15.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated Docs around the requirement of terra-base and aggregate-translations.