+++ packages/terra-form-field/CHANGELOG.md
@@ -3,8 +3,15 @@ ChangeLog
 
 Unreleased
 ----------
+
+3.20.0 - (August 14, 2019)
+------------------
 ### Changed
+* Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts
+
+### Fixed
+* Pass down `isInvalid` to children of Field
 
 3.19.0 - (July 30, 2019)
 ------------------
@@ -34,6 +41,7 @@ Unreleased
 3.14.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated Docs around the requirement of terra-base and aggregate-translations.