+++ packages/terra-image/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+3.15.0 - (August 14, 2019)
+------------------
 ### Changed
 * updated package.json test scripts
 
@@ -20,6 +23,10 @@ Unreleased
 ------------------
 ### Added
 * Added documentation on required peerDependencies
+* Added `Object-fit-images` polyfill for IE.
+* Added `object-fit` property in css.
+* Added prop Fit.
+* Added examples using Fit.
 
 ### Changed
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`