+++ packages/terra-search-field/CHANGELOG.md
@@ -3,7 +3,16 @@ Changelog
 
 Unreleased
 ----------
+
+3.24.0 - (August 14, 2019)
+------------------
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
+
 ### Changed
+* Updated search-field example code to show more usage.
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+* Update jest test snapshot
 * updated package.json test scripts
 
 3.23.0 - (July 30, 2019)
@@ -29,6 +38,7 @@ Unreleased
 3.19.0 - (July 16, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated jest snapshots