+++ packages/terra-section-header/CHANGELOG.md
@@ -3,8 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.20.0 - (August 14, 2019)
+------------------
 ### Changed
 * updated package.json test scripts
+* Remove react and react-dom as dependencies
 
 2.19.0 - (July 30, 2019)
 ------------------